Rename proccomm pkg as pcomm
